shader doPendingMapLoads: col int (not s16) drops counter extsh in block-clear loop (85.51->85.66)

+++ src/main/shader.c
@@ -2770,7 +2770,7 @@ void doPendingMapLoads(void)
     int doLoad;
     int layer;
     int row;
-    s16 col;
+    int col;
     int cell;
     int i;
     int slot;
